backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
sqlalchemy>=2.0.20
alembic>=1.12.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
mcp>=1.4.1
jinja2>=3.1.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.7
cryptography>=41.0.3
pyyaml>=6.0.1
pytest>=7.4.2
httpx>=0.25.0
beautifulsoup4>=4.12.0
markdown>=3.5.1
langgraph>=0.0.19
openai>=1.3.0
fastmcp>=0.2.0
langchain-core>=0.1.0
langchain>=0.0.300
asyncio>=3.4.3
requests>=2.31.0 